Fix state change causing graph to re render
