+++ client/src/Components/Clock.tsx
@@ -4,7 +4,7 @@ import React, {useEffect, useState} from "react";
 
 export default function (props: any) {
     const time = getTimeNow({enabled: true});
-    const [display, setDisplay] = useState<number>(-1)
+    const [display, setDisplay] = useState<number>(-1);
 
     useEffect(() => setDisplay(Math.round((time.getTime() - props.minus) / props.divider)), [time])
